requirements.txt
streamlit==1.38.0
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
torch>=2.2,<3
transformers>=4.41,<5
numpy>=1.26,<2.0
requests>=2.32
PyYAML>=6.0
tqdm>=4.66
